Extract ProximityPrecondition API as interface, prepare to swap
